+++ src/Queues/ServiceBusMessageTopic.cs
@@ -1,9 +1,9 @@
-﻿using PipServices.Components.Auth;
-using PipServices.Commons.Config;
-using PipServices.Components.Connect;
-using PipServices.Commons.Data;
-using PipServices.Commons.Errors;
-using PipServices.Messaging.Queues;
+﻿using PipServices3.Components.Auth;
+using PipServices3.Commons.Config;
+using PipServices3.Components.Connect;
+using PipServices3.Commons.Data;
+using PipServices3.Commons.Errors;
+using PipServices3.Messaging.Queues;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -11,7 +11,7 @@ using System.Threading.Tasks;
 
 using Microsoft.Azure.ServiceBus;
 
-namespace PipServices.Azure.Queues
+namespace PipServices3.Azure.Queues
 {
     /*
     public class ServiceBusMessageTopic : MessageQueue