+++ src/Queues/ServiceBusMessageTopic2.cs
@@ -1,15 +1,15 @@
-﻿using PipServices.Components.Auth;
-using PipServices.Commons.Config;
-using PipServices.Components.Connect;
-using PipServices.Commons.Errors;
-using PipServices.Messaging.Queues;
+﻿using PipServices3.Components.Auth;
+using PipServices3.Commons.Config;
+using PipServices3.Components.Connect;
+using PipServices3.Commons.Errors;
+using PipServices3.Messaging.Queues;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace PipServices.Azure.Queues
+namespace PipServices3.Azure.Queues
 {
     // This implementation doesn't use subscriptions. Don't use it unless you know what you are doing!
     /*