+++ test/Queues/MessageQueueFixture.cs
@@ -1,11 +1,11 @@
-﻿using PipServices.Messaging.Queues;
+﻿using PipServices3.Messaging.Queues;
 
 using System.Threading;
 using System.Threading.Tasks;
 
 using Xunit;
 
-namespace PipServices.Azure.Queues
+namespace PipServices3.Azure.Queues
 {
     public class MessageQueueFixture
     {